+++ ChangeLog.md
@@ -5,6 +5,20 @@
   - A Dockerfile is added to create a Docker image to have a base to start development
     using the DGtal library.(J. Miguel Salazar [#1580](https://github.com/DGtal-team/DGtal/pull/1580)) 
 
+- *Geometry Package*
+
+  - Completes the digital convexity module with new functions
+    related to full convexity: check of full convexity for arbitrary
+    digital sets in nD, and helper classes for using full convexity in
+    practice (local geometric analysis, tangency and shortest paths)
+    (Jacques-Olivier Lachaud,[#1594](https://github.com/DGtal-team/DGtal/pull/1594))
+
+- *Mathematical Package*
+
+   - Add Lagrange polynomials and Lagrange interpolation
+     (Jacques-Olivier Lachaud,[#1594](https://github.com/DGtal-team/DGtal/pull/1594))
+
+
 ## Bug fixes
 - *General*
   - A Dockerfile is added to create a Docker image to have a base to start development